Move executable code to new "bin" directory.
